extracts logic to Queries::CategoryOrganisations
