+++ spec/services/queries/category_organisations_spec.rb
@@ -1,0 +1,6 @@
+require 'spec_helper'
+require_relative '../../../app/services/queries/category_organisations'
+
+
+describe Queries::CategoryOrganisations do
+end
